SLING-7935 - Move launchpad-testing module into the main starter repository
Don't check the version of the scripting.jsp bundle. This check was added
a long time ago, and in the current version of the Starter the bundle version
is recent enough.

Additionally, this pulls in the OSGi core jar as a provided dependency
of the commons.testing jar, and that needs to be declared as a dependency
specifically by all consumers of this test.